Update TSP_AR_MO_Iron_SHUTITDOWN.xml
Alright, I found the error, Varme! It was indeed on my end, and it was a single mistake. When I wrote my Patch, I wrote THINGDEF instead of RECIPEDEF in the XPATH, causing it to fail to replace the results.

Please try again now, I have fixed the problem!

Thank you so much!